tmp constisten KDF API
